Move more definitions into their respective files for iOS to share. - Export shared types from `procedural_filters.ts` - `content_cosmetic` Window definition into `content_cosmetic.ts` - Action types into `content_cosmetic.ts`
