+++ components/definitions/global.d.ts
@@ -15,28 +15,6 @@ type RequestIdleCallbackDeadline = {
 }
 
 declare global {
-  type StyleAction = {
-    type: 'style'
-    arg: string
-  }
-  type RemoveAction = {
-    type: 'remove'
-  }
-  type RemoveAttrAction = {
-    type: 'remove-attr'
-    arg: string
-  }
-  type RemoveClassAction = {
-    type: 'remove-class'
-    arg: string
-  }
-
-  type CosmeticFilterAction = StyleAction | RemoveAction | RemoveAttrAction | RemoveClassAction;
-
-  type ProceduralActionFilter = {
-    selector: ProceduralSelector
-    action?: CosmeticFilterAction
-  }
   interface Window {
     // Typescript doesn't include requestIdleCallback as it's non-standard.
     // Since it's supported in Chromium, we can include it here.
@@ -47,27 +25,5 @@ declare global {
     cancelIdleCallback: ((handle: RequestIdleCallbackHandle) => void)
     alreadyInserted: boolean
     web3: any
-    content_cosmetic: {
-      cosmeticStyleSheet: CSSStyleSheet
-      allSelectorsToRules: Map<string, number>
-      observingHasStarted: boolean
-      hide1pContent: boolean
-      generichide: boolean
-      firstRunQueue: Set<string>
-      secondRunQueue: Set<string>
-      finalRunQueue: Set<string>
-      allQueues: Set<string>[]
-      numQueues: any
-      alreadyUnhiddenSelectors: Set<string>
-      alreadyKnownFirstPartySubtrees: WeakSet
-      _hasDelayOcurred: boolean
-      _startCheckingId: number | undefined
-      firstSelectorsPollingDelayMs: number | undefined
-      switchToSelectorsPollingThreshold : number | undefined
-      fetchNewClassIdRulesThrottlingMs : number | undefined
-      tryScheduleQueuePump: (() => void)
-      proceduralActionFilters?: ProceduralActionFilter[]
-      hasProceduralActions: boolean
-    }
   }
 }
